refactor: move dust app execution and processing in separate function
